`RightSidebar` takes contents as children
