+++ src/components/BackupScreen/BackupScreen.tsx
@@ -1,14 +1,19 @@
 'use client'
 
+import Link from 'next/link'
 import { styled } from 'next-yak'
 import { Suspense } from 'react'
 import { Panel, PanelGroup, PanelResizeHandle } from 'react-resizable-panels'
+import useSWR from 'swr'
 
 import { useStorachaAccount } from '@/hooks/use-plan'
-import { Backup } from '@/types'
+import { formatDate, shortenCID, shortenDID } from '@/lib/ui'
+import { Backup, Snapshot } from '@/types'
+
+import { Box, Center, Heading, Stack, SubHeading, Text } from '../ui'
 
 import { BackupDetail } from './BackupDetail'
-import Sidebar from './RightSidebar'
+import RightSidebar from './RightSidebar'
 
 const BackupContainer = styled.div`
   display: flex;
@@ -34,8 +39,41 @@ const ResizeHandleInner = styled.div`
   }
 `
 
+const Instruction = styled(Text)``
+
+const SnapshotContainer = styled(Stack)`
+  margin-top: 4rem;
+`
+
+const Details = styled(Stack)`
+  margin-top: 4rem;
+`
+
+const DetailName = styled(SubHeading)`
+  color: black;
+`
+
+const DetailValue = styled.div`
+  font-family: var(--font-dm-mono);
+  font-size: 0.75rem;
+`
+
+const SnapshotSummary = styled(Box)`
+  padding: 1rem;
+  font-size: 0.75rem;
+`
+
+const SnapshotLink = styled(Link)`
+  width: 100%;
+  height: 100%;
+`
+
 export const BackupScreen = ({ backup }: { backup?: Backup }) => {
   const account = useStorachaAccount()
+  const { data: snapshots } = useSWR<Snapshot[]>(
+    backup && ['api', `/api/backups/${backup.id}/snapshots`]
+  )
+
   return (
     <BackupContainer>
       <PanelGroup autoSaveId="backup-restore-layout" direction="horizontal">
@@ -51,7 +89,58 @@ export const BackupScreen = ({ backup }: { backup?: Backup }) => {
         </PanelResizeHandle>
         <Panel defaultSize={40} minSize={40}>
           <Suspense>
-            <Sidebar backup={backup} />
+            <RightSidebar>
+              <Heading>Backup & Restore</Heading>
+              {backup ? (
+                <>
+                  <Details $gap="1rem">
+                    <SubHeading>Details</SubHeading>
+                    <Stack $direction="row" $alignItems="center" $gap="1rem">
+                      <DetailName>Account DID</DetailName>
+                      <DetailValue>
+                        {shortenDID(backup.atprotoAccount)}
+                      </DetailValue>
+                    </Stack>
+                    <Stack $direction="row" $alignItems="center" $gap="1rem">
+                      <DetailName>Delegation CID</DetailName>
+                      <DetailValue>
+                        {backup.delegationCid
+                          ? shortenCID(backup.delegationCid)
+                          : 'No delegation set'}
+                      </DetailValue>
+                    </Stack>
+                  </Details>
+                  <SnapshotContainer $gap="1rem">
+                    <SubHeading>Snapshots</SubHeading>
+                    {snapshots?.map((snapshot) => (
+                      <SnapshotSummary
+                        key={snapshot.id}
+                        $background="var(--color-white)"
+                      >
+                        <SnapshotLink href={`/snapshots/${snapshot.id}`}>
+                          <Stack
+                            $direction="row"
+                            $alignItems="center"
+                            $justifyContent="space-between"
+                            $width="100%"
+                          >
+                            <Stack $direction="column" $alignItems="flex-start">
+                              <h3>{formatDate(snapshot.createdAt)} Snapshot</h3>
+                            </Stack>
+                          </Stack>
+                        </SnapshotLink>
+                      </SnapshotSummary>
+                    ))}
+                  </SnapshotContainer>
+                </>
+              ) : (
+                <Center $height="90vh">
+                  <Instruction $fontWeight="600">
+                    Press &quot;Create Backup&quot; to get started!
+                  </Instruction>
+                </Center>
+              )}
+            </RightSidebar>
           </Suspense>
         </Panel>
       </PanelGroup>
